backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-mail==1.4.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.44.1
pandas==2.1.4
numpy<2.0.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
